feat(cli): allow `lagon deploy` & `lagon build` to specify files
