+++ .changeset/smart-hotels-camp.md
@@ -1,0 +1,6 @@
+---
+'@lagon/cli': patch
+'@lagon/docs': patch
+---
+
+Allow `lagon deploy` & `lagon build` to specify files and folders
